refactor: update import paths for prepare_test_rdm_record
