+++ invenio_notify_test/test_endorsement.py
@@ -6,9 +6,9 @@ from sqlalchemy.exc import StatementError
 
 from invenio_notify import proxies, constants
 from invenio_notify.records.models import EndorsementModel
-from invenio_notify_test.conftest import prepare_test_rdm_record
 from invenio_notify_test.fixtures.endorsement_fixture import create_endorsement
 from invenio_notify_test.fixtures.inbox_fixture import create_inbox
+from invenio_notify_test.fixtures.record_fixture import prepare_test_rdm_record
 from invenio_notify_test.fixtures.reviewer_fixture import create_reviewer
